use long

+++ src/cli.h
@@ -2,17 +2,17 @@
 #include <cstdlib>
 
 struct Params {
-    int N = 1000; // number of docs
-    int Q = 100; // number of queries
+    long N = 1000; // number of docs
+    long Q = 100; // number of queries
 };
 
 void parseCommandLine(int argc, char* argv[], Params& params) {
     for (int i = 1; i < argc; ++i) {
         std::string arg = argv[i];
         if (arg == "--N" && i + 1 < argc) {
-            params.N = std::atoi(argv[++i]);
+            params.N = std::atol(argv[++i]);
         } else if (arg == "--Q" && i + 1 < argc) {
-            params.Q = std::atoi(argv[++i]);
+            params.Q = std::atol(argv[++i]);
         }
     }
 }
